src/requirements.txt
web.py
requests
google-api-python-client
google-auth
google-cloud-logging
google-cloud-tasks